\name{CheckBinary}
\alias{CheckBinary}
\title{
  Test Run of X-13ARIMA-SEATS
}
\usage{
  CheckBinary()
}
\description{
  Performs a test run of X-13ARIMA-SEATS. Fails if no output is produced.
}
\author{
  Dirk Eddelbuettel, Christoph Sax

  Maintainer: Dirk Eddelbuettel <edd@debian.org>
}
\examples{
  CheckBinary()
}

