\name{checkX13binary}
\alias{checkX13binary}
\title{
  Test Run of X-13ARIMA-SEATS
}
\usage{
  checkX13binary()
}
\description{
  Performs a test run of X-13ARIMA-SEATS. Fails if no output is produced.
}
\author{
  Dirk Eddelbuettel, Christoph Sax

  Maintainer: Dirk Eddelbuettel <edd@debian.org>
}
\examples{
  checkX13binary()
}

