\name{x13binary-package}
\alias{x13binary-package}
\alias{x13binary}
\docType{package}
\title{
  Provide the \sQuote{X13-ARIMA-SEATS} Seasonal Adjustment Program
}
\description{
  The US Census provides a seasonal adjustment program now
  called 'X-13ARIMA-SEATS' building on both earlier Census programs called
  X11 and X12 as well as the SEATS program by the Bank of Spain. Census
  offers both source and binary versions.

  This package integrates these for use by other R packages.
}
\details{
  Depending on the operating system detected during installation of this
  program, either a binary is downloaded or compiled from sources.
}
\author{
  Dirk Eddelbuettel

  Maintainer: Dirk Eddelbuettel <edd@debian.org>
}
\references{
  https://www.census.gov/srd/www/x13as/
}
\keyword{package}
\examples{
  # Optional simple examples of the most important functions
}
