% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/x13path.R
\name{x13path}
\alias{x13path}
\title{Full Path to X-13ARIMA-SEATS}
\usage{
x13path()
}
\description{
Returns the full full path to the X-13ARIMA-SEATS binary contained in the
package, or \code{""} if the platform is unsupported.
}
\examples{
x13path()
}

