% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkX13binary.R
\name{checkX13binary}
\alias{checkX13binary}
\title{Check if X-13ARIMA-SEATS Runs Properly}
\usage{
checkX13binary(fail.unsupported = FALSE, verbose = TRUE)
}
\arguments{
\item{fail.unsupported}{logical, whether beeing on an unsupported platform
leads to an error.}

\item{verbose}{logical, should a message be returned on success?}
}
\description{
Performs a test run of X-13ARIMA-SEATS. Fails if no output is produced.
}
\examples{
checkX13binary()
}

