% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13binary-package.R
\docType{package}
\name{x13binary-package}
\alias{x13binary-package}
\alias{x13binary}
\title{Provide the \sQuote{X13-ARIMA-SEATS} Seasonal Adjustment Program}
\description{
The US Census provides a seasonal adjustment program now called
 'X-13ARIMA-SEATS' building on both earlier Census programs called X11 and
 X12 as well as the SEATS program by the Bank of Spain. Census offers both
 source and binary versions. This package integrates these for use by other R 
 packages.
}
\references{
\url{https://www.census.gov/data/software/x13as.X-13ARIMA-SEATS.html}
}
\author{
Dirk Eddelbuettel \email{edd@debian.org} and Christoph Sax
}
\keyword{package}
